# Conflicts:
#	SaluteStocksAPI/Properties/PublishProfiles/FTPProfile.pubxml.user
#	SaluteStocksAPI/Service/Background/Loader.cs

+++ SaluteStocksAPI/Models/FundamentalData/Earnings.cs
@@ -13,6 +13,8 @@ public class AnnualEarning
     public string Symbol { get; set; }
     public Earnings Earnings { get; set; }
     
+    
+    
     /// <summary>
     /// years: 2021, 2020, 2019 ... 1996
     /// </summary>
@@ -28,7 +30,7 @@ public class AnnualEarning
     /// привлекательности и эффективности компаний,
     /// действующих на фондовом рынке.
     /// </summary>
-    [JsonProperty("reportedEPS")] public double? ReportedEPS { get; set; }
+    [JsonProperty("reportedEPS")] public decimal? ReportedEPS { get; set; }
 }
 
 public class QuarterlyEarning
@@ -44,13 +46,13 @@ public class QuarterlyEarning
 
     [JsonProperty("reportedDate")] public DateTime? ReportedDate { get; set; }
 
-    [JsonProperty("reportedEPS")] public double? ReportedEPS { get; set; }
+    [JsonProperty("reportedEPS")] public decimal? ReportedEPS { get; set; }
 
-    [JsonProperty("estimatedEPS")] public double? EstimatedEPS { get; set; }
+    [JsonProperty("estimatedEPS")] public decimal? EstimatedEPS { get; set; }
 
-    [JsonProperty("surprise")] public double? Surprise { get; set; }
+    [JsonProperty("surprise")] public decimal? Surprise { get; set; }
 
-    [JsonProperty("surprisePercentage")] public double? SurprisePercentage { get; set; }
+    [JsonProperty("surprisePercentage")] public decimal? SurprisePercentage { get; set; }
 }
 
 public class Earnings : CompanyEntityInfo
